CDK Nags mitigated. Fixed compilation issues
