+++ src/control-plane/index.ts
@@ -3,12 +3,12 @@
 
 export * from './auth/auth-interface';
 export * from './auth/cognito-auth';
+export * from './aws-marketplace/index';
+export * from './billing/index';
 export * from './control-plane-api';
 export * from './control-plane';
-export * from './models/tenant';
-export * from './services';
-export * from './tables';
-export * from './billing/index';
 export * from './ingestor-aggregator/index';
-export * from './tenant-config/tenant-config-service';
-export * from './aws-marketplace/index';
+export * from './models/tenant';
+export * from './tenant-config';
+export * from './tenant-management';
+export * from './user-management';
