+++ src/control-plane/tenant-config/index.ts
@@ -1,0 +1,5 @@
+// Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
+// SPDX-License-Identifier: Apache-2.0
+
+export * from './tenant-config-funcs';
+export * from './tenant-config.service';
