+++ src/control-plane/tenant-management/index.ts
@@ -1,0 +1,6 @@
+// Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
+// SPDX-License-Identifier: Apache-2.0
+
+export * from './tenant-management-funcs';
+export * from './tenant-management.service';
+export * from './tenant-management.table';
